+++ apps/api/src/logistic/catalog/catalog.service.ts
@@ -16,6 +16,13 @@ type GearUpdateForm = CatalogGearForm & {
   id: number;
 };
 
+export type GearLinkedItems = {
+  tasks: number[];
+  actitivities: number[];
+  borrows: number[];
+  purchases: number[];
+};
+
 export class GearNotFoundException extends NotFoundException {
   constructor(id: number) {
     super(`Gear #${id} doesn't exist`);
@@ -70,6 +77,28 @@ export class CatalogService {
   }
 
   async remove(id: number): Promise<void> {
+    const linked = await this.gearRepository.getLinkedItems(id);
+    const hasLinkedItems =
+      linked?.actitivities.length ||
+      linked?.tasks.length ||
+      linked?.borrows.length ||
+      linked?.purchases.length;
+
+    if (hasLinkedItems) {
+      const actitivities = linked.actitivities.map((id) => `FA ${id}`);
+      const tasks = linked.tasks.map((id) => `FT ${id}`);
+      const borrows = linked.borrows.map((id) => `Fiche Emprunt ${id}`);
+      const purchases = linked.purchases.map((id) => `Fiche Achat ${id}`);
+      const allLinkedItems = [
+        ...actitivities,
+        ...tasks,
+        ...borrows,
+        ...purchases,
+      ];
+      const errorMessage = `Impossible de supprimer le matériel, il est lié à : ${allLinkedItems.join(", ")}`;
+      throw new Error(errorMessage);
+    }
+
     return this.gearRepository.removeGear(id);
   }
